@@ -38,52 +38,52 @@ abstract class FormTypeTest extends FormIntegrationTestCase
     {
         parent::setUp();
 
-        $paths = array(
+        $paths = [
             __DIR__ . '/../../../../Resources/views/Form',
             __DIR__ . '/../../../Resources/views',
             (file_exists(VENDOR_DIR . '/symfony/twig-bridge/Resources/views'))
                 ? VENDOR_DIR . '/symfony/twig-bridge/Resources/views'
                 : VENDOR_DIR . '/symfony/twig-bridge/Symfony/Bridge/Twig/Resources/views'
-        );
+        ];
 
         $loader = new StubFilesystemLoader($paths);
 
-        $twig = new \Twig_Environment($loader, array('strict_variables' => true));
+        $twig = new \Twig_Environment($loader, ['strict_variables' => true]);
         $twig->addGlobal('global', '');
         $twig->addExtension(new TranslationExtension(new StubTranslator()));
         $twig->addExtension(new FileTwigExtension());
         $twig->addExtension(new Assets(new FSiFilePathResolver('/adapter/path', 'uploaded')));
-        $rendererEngine = new TwigRendererEngine(array(
+        $rendererEngine = new TwigRendererEngine([
             'form_div_layout.html.twig',
             'Form/form_div_layout.html.twig'
-        ), $twig);
+        ], $twig);
         $renderer = new TwigRenderer(
             $rendererEngine,
-            $this->getMock('\Symfony\Component\Security\Csrf\CsrfTokenManagerInterface')
+            $this->getMockBuilder('Symfony\Component\Security\Csrf\CsrfTokenManagerInterface')->getMock()
         );
 
         if (class_exists('Symfony\Bridge\Twig\Extension\AssetExtension')) {
-            $runtimeLoader = $this->getMock('Twig_RuntimeLoaderInterface');
-            $runtimeLoader->expects($this->any())->method('load')->will($this->returnValueMap(array(
-                array('Symfony\Bridge\Twig\Form\TwigRenderer', $renderer),
-            )));
+            $runtimeLoader = $this->getMockBuilder('Twig_RuntimeLoaderInterface')->getMock();
+            $runtimeLoader->expects($this->any())->method('load')->will($this->returnValueMap([
+                ['Symfony\Bridge\Twig\Form\TwigRenderer', $renderer],
+            ]));
             $twig->addRuntimeLoader($runtimeLoader);
 
             $twig->addExtension(new FormExtension());
             $twig->addExtension(new AssetExtension(
                 new Packages(
-                    new UrlPackage(array('http://local.dev/'), new EmptyVersionStrategy())
+                    new UrlPackage(['http://local.dev/'], new EmptyVersionStrategy())
                 )
             ));
         } else {
             $twig->addExtension(new FormExtension($renderer));
-            $container = $this->getMock('Symfony\Component\DependencyInjection\ContainerInterface');
+            $container = $this->getMockBuilder('Symfony\Component\DependencyInjection\ContainerInterface')->getMock();
             $container->expects($this->any())
                 ->method('get')
                 ->with('templating.helper.assets')
                 ->will($this->returnValue(new LegacyUrlPackage()));
 
-            $request = $this->getMock('Symfony\Component\Routing\RequestContext');
+            $request = $this->getMockBuilder('Symfony\Component\Routing\RequestContext')->getMock();
             $twig->addExtension(new AssetsExtension($container, $request));
         }
 
@@ -102,5 +102,10 @@ abstract class FormTypeTest extends FormIntegrationTestCase
         }
 
         return file_get_contents($path);
+    }
+
+    protected function isSymfony3()
+    {
+        return method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
     }
 }
