@@ -26,9 +26,9 @@ class TranslatableLocaleListener implements EventSubscriberInterface
      */
     public static function getSubscribedEvents()
     {
-        return array(
+        return [
             KernelEvents::REQUEST => 'onKernelRequest'
-        );
+        ];
     }
 
     /**
