@@ -32,9 +32,9 @@ class UploadablePostLoadListener implements EventSubscriber
      */
     public function getSubscribedEvents()
     {
-        return array(
+        return [
             'postLoad'
-        );
+        ];
     }
 
     /**
