@@ -14,43 +14,36 @@ use FSi\Bundle\DoctrineExtensionsBundle\Tests\Fixtures\Form\Extension\FSiFileExt
 
 class RemovableFileTypeTest extends FormTypeTest
 {
-    /**
-     * @return \FSi\Bundle\DoctrineExtensionsBundle\Tests\Fixtures\Form\Extension\FSiFileExtension
-     */
     public function getExtensions()
     {
-        return array(
+        return [
             new FSiFileExtension()
-        );
+        ];
     }
 
     public function testFormRendering()
     {
-        $self = $this;
-
         $file = $this->getMockBuilder('FSi\DoctrineExtensions\Uploadable\File')
             ->disableOriginalConstructor()
             ->getMock();
 
         $file->expects($this->any())
             ->method('getFilesystem')
-            ->will($this->returnCallback(function() use ($self) {
-                $fileSystem = $self->getMockBuilder('FSi\Bundle\DoctrineExtensionsBundle\Listener\Uploadable\Filesystem')
+            ->will($this->returnCallback(function() {
+                $fileSystem = $this->getMockBuilder('FSi\Bundle\DoctrineExtensionsBundle\Listener\Uploadable\Filesystem')
                     ->disableOriginalConstructor()
                     ->getMock();
 
-                $fileSystem->expects($self->any())
+                $fileSystem->expects($this->any())
                     ->method('getBaseUrl')
-                    ->will($self->returnValue('/uploaded/'));
+                    ->will($this->returnValue('/uploaded/'));
 
-                $fileSystem->expects($self->any())
+                $fileSystem->expects($this->any())
                     ->method('getAdapter')
-                    ->will($self->returnCallback(function() use ($self) {
-                        $adapter = $self->getMockBuilder('Gaufrette\Adapter\Local')
+                    ->will($this->returnCallback(function() {
+                        return $this->getMockBuilder('Gaufrette\Adapter\Local')
                             ->disableOriginalConstructor()
                             ->getMock();
-
-                        return $adapter;
                     }));
 
                 return $fileSystem;
@@ -74,14 +67,9 @@ class RemovableFileTypeTest extends FormTypeTest
 
         $form->add('file', $this->isSymfony3() ? 'FSi\Bundle\DoctrineExtensionsBundle\Form\Type\FSi\RemovableFileType' : 'fsi_removable_file');
 
-        $html = $this->twig->render('form_with_fsi_file.html.twig', array(
+        $html = $this->twig->render('form_with_fsi_file.html.twig', [
             'form' => $form->createView()
-        ));
+        ]);
         $this->assertSame($this->getExpectedHtml('form_with_fsi_removable_file.html'), $html);
-    }
-
-    private function isSymfony3()
-    {
-        return method_exists('Symfony\Component\Form\AbstractType', 'getBlockPrefix');
     }
 }
